move navbar into app components
