+++ apps/frontend/src/features/app/components/Navbar.vue
