+++ apps/frontend/src/features/app/components/__tests__/Navbar.spec.ts
